python_backend/convert_ckpt_to_onnx.py
"""
将 LaMa PyTorch checkpoint 转换为 ONNX 格式
"""
import torch
import torch.nn as nn
import onnx
from onnx import shape_inference
import sys
import os

def convert_lama_to_onnx(ckpt_path, output_path="lama_model.onnx"):
    """
    转换 LaMa checkpoint 到 ONNX
    
    注意：这个脚本需要 LaMa 的模型定义代码
    由于我们没有完整的模型定义，这里提供一个简化的转换流程
    """
    print(f"📥 加载 checkpoint: {ckpt_path}")
    
    try:
        # 加载 checkpoint
        checkpoint = torch.load(ckpt_path, map_location='cpu')
        
        print("📦 Checkpoint 内容:")
        if isinstance(checkpoint, dict):
            for key in checkpoint.keys():
                print(f"  - {key}")
        
        # 这里需要 LaMa 的模型定义
        # 由于我们没有完整的模型代码，建议使用已转换好的 ONNX 模型
        
        print("\n⚠️  转换 LaMa checkpoint 需要完整的模型定义代码")
        print("💡 建议使用以下方案之一：")
        print("\n方案 1: 使用 lama-cleaner 项目的预转换模型")
        print("  下载地址: https://github.com/Sanster/models/releases")
        print("\n方案 2: 使用简化的 ONNX 模型")
        print("  我们可以使用一个轻量级的修复模型")
        
    except Exception as e:
        print(f"❌ 加载失败: {e}")
        return False
    
    return False

if __name__ == "__main__":
    if len(sys.argv) < 2:
        print("用法: python convert_ckpt_to_onnx.py <ckpt文件路径>")
        sys.exit(1)
    
    ckpt_path = sys.argv[1]
    if not os.path.exists(ckpt_path):
        print(f"❌ 文件不存在: {ckpt_path}")
        sys.exit(1)
    
    convert_lama_to_onnx(ckpt_path)
